lightview off by default, enabled for tablets,desktops
Instead of having two instances of each primarysource, and
showing/hiding them using CSS and media queries,
have a single instance of each one,
use jQuery and Modernizr to enable lightview when page is viewed
on a tablet or desktop.

At this point, jQuery/Modernizr are correctly detecting tablet
size and up.
refs #196