fix formatting on editor's message
The editor's message looks fine on mobile devices, where it is
displayed in its own page.  But on tablets/desktops where it's
displayed in a lightbox the embedded page was always larger than
the lightbox.
This change moves the text to its own document and creates
enclosure templates for the ownpage and lightboxed page, with
appropriate html/bootstrap wrapping.
refs #209