use local copy of OpenLayers; rm Modernizr centering (works on IE9)
Maps weren't working on IE9.  Looks like the problem was a bit of
Javascript that used Modernizr to determine the browser width
and adjust the map center accordingly.
Also, downloaded OpenLayers so we can have our own copy rather
than whatever possibly broken version is currently on the
OpenLayers site.