start work on sorting page titles by sortkey
Pages for people should be listed by their surname.
Those pages are created with a MediaWiki sortkey.
It's possible to retrieve lists of pages arranged by sortkey,
and possible to get the sortkey in results.
But higher-layer functions are not used to having that sortkey
and don't know what to do with it yet.
refs #225