always remove_status_markers regardless of debug status
Previous behavior was to show "This page is complete and will be
published in the public encyclopedia" when in DEBUG mode.
This was appropriate when the app connected with the back-end
MediaWiki whenever page was requested, but doesn't work now that
pages are pre-rendered and stored in Elasticsearch.
This change always removes those alert boxes (and categories?).