ddr,templates: Use DDR object img_url for lightbox
Lightbox is that Javascript thing that displays the image in a
popup when you click on the thumbnail.  Thumbnails are generated
using object.img_thumb because sorl.thumbnail can't get past
CloudFlare.  But lightbox can just use the regular URL.
Previously the code was trying to construct a URL but this was not
working and users were seeing 403 errors in the popup.